// src/seaportEx/types.ts
export type InputCriteria = {
    identifier: string;
    proof: string[];
};

export type NftItemType =
    | ItemType.ERC721
    | ItemType.ERC1155
    | ItemType.ERC721_WITH_CRITERIA
    | ItemType.ERC1155_WITH_CRITERIA;

export enum BasicOrderRouteType {
    ETH_TO_ERC721,
    ETH_TO_ERC1155,
    ERC20_TO_ERC721,
    ERC20_TO_ERC1155,
    ERC721_TO_ERC20,
    ERC1155_TO_ERC20,
}

export type OfferItem = {
    itemType: ItemType;
    token: string;
    identifierOrCriteria: string;
    startAmount: string;
    endAmount: string;
};

export type ConsiderationItem = {
    itemType: ItemType;
    token: string;
    identifierOrCriteria: string;
    startAmount: string;
    endAmount: string;
    recipient: string;
};

export type Item = OfferItem | ConsiderationItem;

export type OrderParameters = {
    offerer: string;
    zone: string;
    orderType: OrderType;
    startTime: BigNumberish;
    endTime: BigNumberish;
    zoneHash: string;
    salt: string;
    offer: OfferItem[];
    consideration: ConsiderationItem[];
    totalOriginalConsiderationItems: BigNumberish;
    conduitKey: string;
};

export type OrderComponents = OrderParameters & { counter: number };

export type Order = {
    parameters: OrderParameters;
    signature: string;
};

export type OrderStatus = {
    isValidated: boolean;
    isCancelled: boolean;
    totalFilled: BigNumber;
    totalSize: BigNumber;
};